+++ src/main/java/com/podium/service/GenderService.java
@@ -1,11 +1,15 @@
 package com.podium.service;
 
-import com.podium.model.dto.request.gender.GenderRequestDto;
-import com.podium.model.dto.response.gender.GenderResponseDto;
+import com.podium.model.dto.request.GenderRequestDto;
+import com.podium.model.dto.response.GenderResponseDto;
+import com.podium.model.entity.Discipline;
 import com.podium.model.entity.Gender;
 import com.podium.repository.GenderRepository;
+import com.podium.service.exception.PodiumEntityAlreadyExistException;
+import com.podium.service.exception.PodiumEntityNotFoundException;
 import org.springframework.stereotype.Service;
 
+import javax.transaction.Transactional;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -24,8 +28,12 @@ public class GenderService {
 
     public GenderResponseDto findByGenderName(String genderName){
 
-       return this.convertEntityToResponseDto(
-               this.genderRepository.findByGender(genderName));
+        Gender gender = this.genderRepository
+                .findByGender(genderName)
+                .orElseThrow(() -> new PodiumEntityNotFoundException("Gender"));
+
+        return this.convertEntityToResponseDto(gender);
+
     }
 
     public Iterable<GenderResponseDto> findAllGenders(){
@@ -41,11 +49,21 @@ public class GenderService {
         return responseDtos;
     }
 
+    @Transactional
     public void addGender(GenderRequestDto requestDto){
+
+        if(this.genderRepository.existsByGender(requestDto.getGender()))
+            throw new PodiumEntityAlreadyExistException("Gender");
+
         this.genderRepository.save(this.convertRequestDtoToEntity(requestDto));
     }
 
+    @Transactional
     public void deleteGenderByName(String name){
+
+        if(!this.genderRepository.existsByGender(name))
+            throw new PodiumEntityNotFoundException("Gender");
+
         this.genderRepository.deleteByGender(name);
     }
